Move all 3D_SW Joints into joints_3d_sw
