+++ servers/physics_3d/SCsub
@@ -3,5 +3,3 @@
 Import("env")
 
 env.add_source_files(env.servers_sources, "*.cpp")
-
-SConscript("joints/SCsub")
