--- servers/physics_3d/joints/SCsub
@@ -1,5 +1,0 @@
-#!/usr/bin/env python
-
-Import("env")
-
-env.add_source_files(env.servers_sources, "*.cpp")
